rpc reconnects on error
To enable reconnect functionality:
````
  rpc.reconnect = true;
  rpc.reconnectDelay = 5000; // no. of milli-seconds to wait before attempting reconnect
````
To dynamically set the reconnct delay, or control the number of reconnect attempts, handle the rpc error as below:
````
let nAttempts = 0;
rpc.on('error', err => {
    rpc.reconnect = nAttempts < 5;
    rpc.reconnectDelay = ++nAttempts * 1000;
})
````
The above example code does reconnect attempts a maximum of 5 times, with increasing time delay of 1000, 2000, 3000... 

Calling `Close()` on the client resets the `reconnect` to be false. If you connect it again, then you have to manually set the `reconnect` option, if required.